Implement target Values interface, modify DataEntryPage and TargetValues page. Hide selectors &  get form from an specific project,dataSet & periodicity
