+++ src/pages/root/Root.tsx
@@ -4,7 +4,7 @@ import LandingPage from "../landing-page/LandingPage";
 import ProjectsList from "../projects-list/ProjectsList";
 import Report from "../report/Report";
 import DataEntry from "../data-entry/DataEntry";
-import TargetValues from "../target-values/TargetValues";
+import TargetValuesPage from "../target-values/TargetValuesPage";
 import Dashboard from "../dashboard/Dashboard";
 import { generateUrl } from "../../router";
 import ProjectWizard from "../project-wizard/ProjectWizard";
@@ -17,7 +17,7 @@ const Root = () => {
                 <Route path={generateUrl("projects")} render={() => <ProjectsList />} />
                 <Route path={generateUrl("report")} render={() => <Report />} />
                 <Route path={generateUrl("dataEntry")} render={() => <DataEntry />} />
-                <Route path={generateUrl("targetValues")} render={() => <TargetValues />} />
+                <Route path={generateUrl("targetValues")} render={() => <TargetValuesPage />} />
                 <Route path={generateUrl("dashboard")} render={() => <Dashboard />} />
                 <Route render={() => <LandingPage />} />
             </Switch>
