+++ src/pages/target-values/TargetValues.tsx
@@ -1,9 +1,4 @@
 import React, { useEffect, useState } from "react";
-import i18n from "../../locales";
-import PageHeader from "../../components/page-header/PageHeader";
-// import { Button } from "@material-ui/core";
-import { useHistory } from "react-router";
-import { History } from "history";
 //@ts-ignore
 import { useConfig } from "@dhis2/app-runtime";
 
@@ -29,37 +24,42 @@ function waitforElementToLoad(iframeDocument: any, selector: string) {
         check();
     });
 }
-function goTo(history: History, url: string) {
-    history.push(url);
-}
+const setEntryStyling = async (iframe: any) => {
+    const iframeDocument = iframe.contentWindow.document;
 
-const TargetValues: React.FC = () => {
-    const title = i18n.t("Set Target Values for Project");
-    const subtitle = i18n.t(`This is just an example of a description`);
-    const stylesSubtitle = { marginBottom: 10, marginLeft: 15 };
+    await waitforElementToLoad(iframeDocument, "#selectedDataSetId");
+
+    iframeDocument.querySelector("#currentSelection").remove();
+    iframeDocument.querySelector("#header").remove();
+    iframeDocument.querySelector("html").style.overflow = "hidden";
+    iframeDocument.querySelector("#leftBar").style.display = "none";
+    iframeDocument.querySelector("#selectionBox").style.display = "none";
+    iframeDocument.querySelector("body").style.marginTop = "-55px";
+
+    autoResizeIframeByContent(iframe);
+
+    //selecting the form that we want
+
+    const dataSetSelector = iframeDocument.querySelector("#selectedDataSetId");
+    const periodSelector = iframeDocument.querySelector("#selectedPeriodId");
+
+    await waitforElementToLoad(dataSetSelector, "option");
+    iframeDocument.querySelector("#moduleHeader").remove();
+    dataSetSelector.value = "BfMAe6Itzgt";
+    dataSetSelector.onchange();
+
+    await waitforElementToLoad(periodSelector, "option");
+    periodSelector.value = "201910";
+
+    // getting periodSelector options and select it
+    setTimeout(() => periodSelector.onchange(), 10);
+};
+
+const TargetValues = () => {
     const [loading, setLoading] = useState(false);
     const { baseUrl } = useConfig();
     const iFrameSrc = `${baseUrl}/dhis-web-dataentry/index.action`;
-
     const iframeRef: React.RefObject<HTMLIFrameElement> = React.createRef();
-    const setEntryStyling = async (iframe: any) => {
-        const iframeDocument = iframe.contentWindow.document;
-
-        await waitforElementToLoad(iframeDocument, "#selectedDataSetId");
-        iframeDocument.querySelector("#header").remove();
-        iframeDocument.querySelector("#actions").remove();
-        iframeDocument.querySelector("#moduleHeader").remove();
-        // iframeDocument.querySelector("#leftBar").style.display = "none";
-        iframeDocument.querySelector("html").style.overflow = "hidden";
-
-        iframeDocument.querySelector("body").style.marginTop = "-55px";
-        autoResizeIframeByContent(iframe);
-    };
-    const help = i18n.t(
-        `This is just an example of a help instruction: Register your target values for this project`
-    );
-    const history = useHistory();
-    const goBack = () => goTo(history, "/projects");
 
     useEffect(() => {
         const iframe = iframeRef.current;
@@ -71,18 +71,7 @@ const TargetValues: React.FC = () => {
 
     return (
         <React.Fragment>
-            <PageHeader title={title} help={help} onBackClick={goBack} />
-            <div style={stylesSubtitle}>{subtitle}</div>
-            <div style={styles.backgroundIframe}>
-                <iframe
-                    ref={iframeRef}
-                    title="Target Values"
-                    src={iFrameSrc}
-                    style={styles.iframe}
-                ></iframe>
-                {/* <Button autoFocus>{i18n.t("Close")}</Button>
-                <Button>{i18n.t("Save")}</Button> */}
-            </div>
+            <iframe ref={iframeRef} src={iFrameSrc} style={styles.iframe}></iframe>
         </React.Fragment>
     );
 };
